fix smufl models
