--- src/models/smufl/core/rest.ts
@@ -1,5 +1,0 @@
-import * as Core from ".";
-
-export class Rest {
-  constructor(element: ConstructorParameters<typeof Core.Element>[0]) {}
-}
