--- src/models/smufl/point.ts
@@ -1,4 +1,0 @@
-export class Point {
-  x = 0;
-  y = 0;
-}
