--- src/models/smufl/rect.ts
@@ -1,5 +1,0 @@
-import * as SMUFL from ".";
-export class Rect extends SMUFL.Point {
-  width = 0;
-  height = 0;
-}
