--- src/models/smufl/unit2x/index.ts
@@ -1,1 +1,0 @@
-export * from "./pitch";
