--- src/models/tsmufl/bpm.ts
@@ -1,2 +1,0 @@
-import * as Core from "../core";
-export class Bpm extends Core.Bpm {}
