--- src/models/tsmufl/clef.ts
@@ -1,6 +1,0 @@
-import * as SMUFL from ".";
-export class Clef {
-  sign: "F" | "G" = "F";
-  line: number = 2;
-  glyph = new SMUFL.Glyph("gClef");
-}
