--- src/models/tsmufl/layout.ts
@@ -1,1 +1,0 @@
-export class Layout {}
