--- src/models/tsmufl/part.ts
@@ -1,7 +1,0 @@
-import * as SMUFL from "smufl";
-export class Part {
-  tracks;
-  constructor({ tracks }: { tracks: SMUFL.Track[] }) {
-    this.tracks = tracks;
-  }
-}
