--- src/models/tsmufl/rect.ts
@@ -1,4 +1,0 @@
-export class Rect {
-  width = 0;
-  height = 0;
-}
