--- src/models/tsmufl/row.ts
@@ -1,1 +1,0 @@
-export class Row {}
