--- src/models/tsmufl/staff.ts
@@ -1,4 +1,0 @@
-import * as SMUFL from ".";
-export class Staff {
-  staffLines = new SMUFL.StaffLines();
-}
